Move mixed rate ops get rate into utils
