+++ lib/framestamp_range.ex
@@ -24,9 +24,9 @@ defmodule Vtc.Framestamp.Range do
 
   alias Vtc.Ecto.Postgres.PgFramestamp
   alias Vtc.Framestamp
-  alias Vtc.Framestamp.MixedRateArithmeticError
   alias Vtc.Framestamp.Range.MixedOutTypeArithmeticError
   alias Vtc.Source.Frames
+  alias Vtc.Utils.MixedRateOps
 
   @typedoc """
   Whether the end point should be treated as the Range's boundary (:exclusive), or its
@@ -565,7 +565,7 @@ defmodule Vtc.Framestamp.Range do
     inherit_out_type = Keyword.get(opts, :inherit_out_type, false)
     out_type = get_mixed_out_type([a, b], inherit_out_type, func_name)
 
-    case MixedRateArithmeticError.get_rate(a.in, b.in, inherit_rate, func_name) do
+    case MixedRateOps.get_rate(a.in, b.in, inherit_rate, func_name) do
       {:ok, framerate} ->
         zero_framestamp = Framestamp.with_frames!(0, framerate)
         zero_range = with_duration!(zero_framestamp, zero_framestamp)
@@ -592,7 +592,7 @@ defmodule Vtc.Framestamp.Range do
     result =
       calc_with_exclusive([a, b], inherit_out_type, func_name, fn a, b ->
         if do_calc?.(a, b) do
-          case MixedRateArithmeticError.get_rate(a.in, b.in, inherit_rate, func_name) do
+          case MixedRateOps.get_rate(a.in, b.in, inherit_rate, func_name) do
             {:ok, new_rate} ->
               overlap_in = Enum.max([a.in, b.in], Framestamp)
               overlap_in = Framestamp.with_seconds!(overlap_in.seconds, new_rate)
